de aqui partiremos para implementar JWT

+++ src/main/java/com/security/springsecurity/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package com.security.springsecurity.repository;
 
-import com.security.springsecurity.entity.User;
+import com.security.springsecurity.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
